Refactor localization processing layout
