+++ app/__init__.py
@@ -1,0 +1,3 @@
+"""MCModLocalizer application package."""
+
+__all__ = []
