+++ app/core/__init__.py
@@ -1,0 +1,13 @@
+"""Core utilities for MCModLocalizer."""
+
+from .constants import SYSTEM_INSTRUCTIONS_BASE, USER_TEMPLATE
+from .token_protection import protect_tokens, restore_tokens
+from .usage import UsageStats
+
+__all__ = [
+    "SYSTEM_INSTRUCTIONS_BASE",
+    "USER_TEMPLATE",
+    "protect_tokens",
+    "restore_tokens",
+    "UsageStats",
+]
