+++ app/core/json_io.py
@@ -1,0 +1,23 @@
+"""Helpers for reading and writing JSON files."""
+from __future__ import annotations
+
+import json
+from pathlib import Path
+from typing import Dict
+
+
+def load_json(path: Path) -> Dict[str, str]:
+    if not path.exists():
+        return {}
+    with path.open("r", encoding="utf-8") as f:
+        return json.load(f)
+
+
+def write_json(path: Path, data: Dict[str, str]) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    tmp = path.with_suffix(".tmp.json")
+    tmp.write_text(json.dumps(data, ensure_ascii=False, indent=2), encoding="utf-8")
+    tmp.replace(path)
+
+
+__all__ = ["load_json", "write_json"]
