+++ app/main.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import flet as ft
 
-from ui import main as app_main
+from .ui import main as app_main
 
 
 if __name__ == "__main__":
