+++ app/ui/__init__.py
@@ -1,0 +1,5 @@
+"""UI package for MCModLocalizer."""
+
+from .app import LocalizeApp, TranslationSummary, main
+
+__all__ = ["LocalizeApp", "TranslationSummary", "main"]
