+++ app/ui/app.py
@@ -16,10 +16,10 @@ from xml.sax import saxutils
 
 import flet as ft
 
-from processing import ExtractionResult, extract_localizations, translate_localizations
+from ..services import ExtractionResult, extract_localizations, translate_localizations
 
 APP_NAME = "MC Localizer"
-BASE_DIR = Path(__file__).resolve().parent
+BASE_DIR = Path(__file__).resolve().parent.parent
 RESOURCE_TEMPLATE_DIR = BASE_DIR / "a"
 
 try:
